% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSurvFigs.R
\name{combineSurvFigs}
\alias{combineSurvFigs}
\title{Combine components of a survival figure}
\usage{
combineSurvFigs(KMFig, NumRiskFig, heights = c(2.5, 1))
}
\arguments{
\item{KMFig}{A ggplot2 object that serves as the Kaplan-Meier figure}

\item{NumRiskFig}{A ggplot2 object that serves as the number at risk chart}

\item{heights}{A vector of two numeric values that indicate the ratio of the first row to the second}
}
\value{
A grid figure
}
\description{
Create a grid of survival ggplot2 figures
}
\details{
You likely won't use this function unless you are modifying the individual objects that
are returned from \code{makeKMFig} and \code{makeNumRiskFig}
}
\examples{

topFigure <- makeKMFig(
  formula = survival::Surv(stime, scens) ~ treatment,
  data = mice, ribbon = TRUE
)
bottomChart <- makeNumRiskFig(
  formula = survival::Surv(stime, scens) ~ treatment,
  data = mice
)
combineSurvFigs(topFigure, bottomChart)
combineSurvFigs(topFigure, bottomChart, heights = c(4, 1))
combineSurvFigs(topFigure, bottomChart, heights = c(1, 0.4))
}
