% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKMdf.R
\name{makeKMdf}
\alias{makeKMdf}
\title{Make a data frame for a Kaplan-Meier curve}
\usage{
makeKMdf(formula, data)
}
\arguments{
\item{formula}{A formula object, which must have a `Surv` object as the response on the left of the `~` operator and the grouping term (often the treatment) on the right.}

\item{data}{A data frame in which to interpret the variables named in the formula}
}
\value{
A data frame with columns group, time, n.risk, n.event, n.censor, surv, low, and upp
}
\description{
Create a data frame for plotting a Kaplan-Meier curve by pulling components from \code{\link[survival:survfit.formula]{survival:survfit.formula}}
}
\examples{
makeKMdf(formula = survival::Surv(stime, scens) ~ treatment, data = mice)
}
