% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKMFig.R
\name{makeKMFig}
\alias{makeKMFig}
\title{Make a survival curve figure as a ggplot2 object}
\usage{
makeKMFig(formula, data, timepts = NULL, colors = NULL, ribbon = TRUE)
}
\arguments{
\item{formula}{A formula object, which must have a `Surv` object as the response on the left of the `~` operator and the grouping term (often the treatment) on the right.}

\item{data}{A data frame in which to interpret the variables named in the formula}

\item{timepts}{A numeric vector of time points to show on the x-axis}

\item{colors}{Custom colors for each group line}

\item{ribbon}{Logical to show a semi-transparent confidence band ribbon}
}
\value{
A ggplot2 object of a Kaplan-Meier curve
}
\description{
Make a survival curve figure as a ggplot2 object
}
\examples{
makeKMFig(formula = survival::Surv(stime, scens) ~ treatment, data = mice)

makeKMFig(formula = survival::Surv(stime, scens) ~ treatment, data = mice) +
  ggplot2::labs(y = "OS", color = "Treatment")

makeKMFig(formula = survival::Surv(stime, scens) ~ treatment, data = mice, ribbon = FALSE) +
  ggplot2::scale_color_brewer(palette = "Set1") +
  ggplot2::scale_fill_brewer(palette = "Set1")
}
