% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumRiskFig.R
\name{makeNumRiskFig}
\alias{makeNumRiskFig}
\title{Make a chart of number at risk as a ggplot2 figure}
\usage{
makeNumRiskFig(formula, data, timepts, colors)
}
\arguments{
\item{formula}{A formula object, which must have a `Surv` object as the response on the left of the `~` operator and the grouping term (often the treatment) on the right.}

\item{data}{A data frame in which to interpret the variables named in the formula}

\item{timepts}{A numeric vector of time points to show on the x-axis}

\item{colors}{Custom colors for each group line}
}
\value{
A ggplot2 object number at risk chart
}
\description{
Make a chart of number at risk as a ggplot2 figure
}
\examples{
makeNumRiskFig(formula = survival::Surv(stime, scens) ~ treatment, data = mice)

makeNumRiskFig(formula = survival::Surv(stime, scens) ~ treatment, data = mice) +
  ggplot2::labs(x = "Patients at risk")

makeNumRiskFig(
  formula = survival::Surv(stime, scens) ~ treatment,
  data = mice,
  timepts = c(30, 45, 60, 75, 90)
) +
  ggplot2::scale_color_brewer(palette = "Set1")

makeNumRiskFig(
  formula = survival::Surv(stime, scens) ~ treatment,
  data = mice,
  timepts = c(30, 45, 60, 75, 90),
  colors = c("black", "blue", "red", "green")
)
}
