% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickSurvFig.R
\name{quickSurvFig}
\alias{quickSurvFig}
\title{Create a quick Kaplan-Meier figure with number-at-risk chart}
\usage{
quickSurvFig(
  formula,
  data,
  timepts = NULL,
  colors = NULL,
  ribbon = TRUE,
  heights = c(2.5, 1)
)
}
\arguments{
\item{formula}{A formula object, which must have a `Surv` object as the response on the left of the `~` operator and the grouping term (often the treatment) on the right.}

\item{data}{A data frame in which to interpret the variables named in the formula}

\item{timepts}{A numeric vector of time points to show on the x-axis}

\item{colors}{Custom colors for each group line}

\item{ribbon}{Logical to show a semi-transparent confidence band ribbon}

\item{heights}{A vector of two numeric values that indicate the ratio of the first row to the second}
}
\value{
A grid figure
}
\description{
Create a grid of survival ggplot2 figures quickly, with little ability to modify individual components
}
\examples{
quickSurvFig(survival::Surv(stime, scens) ~ treatment, data = mice)
quickSurvFig(survival::Surv(stime, scens) ~ treatment,
  data = mice,
  timepts = c(30, 45, 60, 75, 90),
  colors = c("black", "blue", "red", "green"),
  ribbon = FALSE,
  heights = c(4, 1)
)
}
